Add condition cards component
Close #37
Create multiple portraits because it's not possible to paint images
using only script library.
Children portraits need to be updated before every paintFront()
execution, it can't be updated inside that function. Also binding a
function to the parent portrait is too complex, to make it simple add a
button to run the updates.